+++ Not Kingdom Rush/Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
 public interface IDamageable 
-{
-    void TakeDamage(float amount);
+{  
+    void TakeDamage(int amount);
 }
